requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
chromadb==1.0.20
click==8.2.1
coloredlogs==15.0.1
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
durationpy==0.10
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.8
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mysqlclient==2.2.7
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.3
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pd==0.0.4
playwright==1.55.0
portalocker==3.2.0
posthog==5.4.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
selenium==4.35.0
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
tokenizers==0.21.4
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
